+++ pm4py/objects/bpmn/exporter/variants/etree.py
@@ -1,18 +1,23 @@
 '''
-    This file is part of PM4Py (More Info: https://pm4py.fit.fraunhofer.de).
+    PM4Py – A Process Mining Library for Python
+Copyright (C) 2024 Process Intelligence Solutions UG (haftungsbeschränkt)
 
-    PM4Py is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as
+published by the Free Software Foundation, either version 3 of the
+License, or any later version.
 
-    PM4Py is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Affero General Public License for more details.
 
-    You should have received a copy of the GNU General Public License
-    along with PM4Py.  If not, see <https://www.gnu.org/licenses/>.
+You should have received a copy of the GNU Affero General Public License
+along with this program.  If not, see this software project's root or
+visit <https://www.gnu.org/licenses/>.
+
+Website: https://processintelligence.solutions
+Contact: info@processintelligence.solutions
 '''
 import uuid
 
@@ -73,7 +78,8 @@ def get_xml_string(bpmn_graph, parameters=None):
     all_processes = set()
     process_planes = {}
     process_process = {}
-    process_participants = {}
+    process_collaboration = None
+
     for node in bpmn_graph.get_nodes():
         all_processes.add(node.get_process())
     for flow in bpmn_graph.get_flows():
@@ -81,21 +87,28 @@ def get_xml_string(bpmn_graph, parameters=None):
 
     if len(all_processes) > 1:
         # when several swimlanes exist, their elements should be annexed to the same BPMN plane
-        bpmn_plane_id = "id" + str(uuid.uuid4())
+        collaboration_nodes = [x for x in bpmn_graph.get_nodes() if isinstance(x, BPMN.Collaboration)]
+        if collaboration_nodes:
+            bpmn_plane_id = collaboration_nodes[0].get_id()
+
         process_collaboration = ET.SubElement(definitions, "bpmn:collaboration", {"id": bpmn_plane_id})
 
-        for process in all_processes:
-            part_id = "id" + str(uuid.uuid4())
-            ET.SubElement(process_collaboration, "bpmn:participant", {"id": part_id, "name": process, "processRef": "id" + process })
-            process_participants[process] = part_id
+        participant_nodes = [x for x in bpmn_graph.get_nodes() if isinstance(x, BPMN.Participant)]
+        if participant_nodes:
+            for process in participant_nodes:
+                ET.SubElement(process_collaboration, "bpmn:participant", {"id": process.get_id(), "name": process.get_name(), "processRef": "id" + process.process_ref})
+                all_processes.add(process.process_ref)
     else:
         bpmn_plane_id = "id" + list(all_processes)[0]
 
     for process in all_processes:
-        p = ET.SubElement(definitions, "bpmn:process",
-                          {"id": "id" + process, "isClosed": "false", "isExecutable": "false",
-                           "processType": "None"})
-        process_process[process] = p
+        if process != bpmn_plane_id:
+            p = ET.SubElement(definitions, "bpmn:process",
+                              {"id": "id" + process, "isClosed": "false", "isExecutable": "false",
+                               "processType": "None"})
+            process_process[process] = p
+        elif process_collaboration is not None:
+            process_process[process] = process_collaboration
 
     diagram = ET.SubElement(definitions, "bpmndi:BPMNDiagram", {"id": "id" + str(uuid.uuid4()), "name": "diagram"})
 
@@ -108,12 +121,13 @@ def get_xml_string(bpmn_graph, parameters=None):
         for node in bpmn_graph.get_nodes():
             process = node.get_process()
 
-            node_shape = ET.SubElement(process_planes[process], "bpmndi:BPMNShape",
-                                       {"bpmnElement": node.get_id(), "id": node.get_id() + "_gui"})
-            node_shape_layout = ET.SubElement(node_shape, "omgdc:Bounds",
-                                              {"height": str(layout.get(node).get_height()), "width": str(layout.get(node).get_width()),
-                                               "x": str(layout.get(node).get_x()),
-                                               "y": str(layout.get(node).get_y())})
+            if process != bpmn_plane_id:
+                node_shape = ET.SubElement(process_planes[process], "bpmndi:BPMNShape",
+                                           {"bpmnElement": node.get_id(), "id": node.get_id() + "_gui"})
+                node_shape_layout = ET.SubElement(node_shape, "omgdc:Bounds",
+                                                  {"height": str(layout.get(node).get_height()), "width": str(layout.get(node).get_width()),
+                                                   "x": str(layout.get(node).get_x()),
+                                                   "y": str(layout.get(node).get_y())})
 
         for flow in bpmn_graph.get_flows():
             process = flow.get_process()
@@ -127,7 +141,11 @@ def get_xml_string(bpmn_graph, parameters=None):
     for node in bpmn_graph.get_nodes():
         process = process_process[node.get_process()]
 
-        if isinstance(node, BPMN.StartEvent):
+        if isinstance(node, BPMN.TextAnnotation):
+            annotation = ET.SubElement(process, "bpmn:textAnnotation", {"id": node.get_id()})
+            text = ET.SubElement(annotation, "bpmn:text")
+            text.text = node.text
+        elif isinstance(node, BPMN.StartEvent):
             isInterrupting = "true" if node.get_isInterrupting() else "false"
             parallelMultiple = "true" if node.get_parallelMultiple() else "false"
             task = ET.SubElement(process, "bpmn:startEvent",
@@ -137,28 +155,36 @@ def get_xml_string(bpmn_graph, parameters=None):
             task = ET.SubElement(process, "bpmn:endEvent", {"id": node.get_id(), "name": node.get_name()})
         elif isinstance(node, BPMN.IntermediateCatchEvent):
             task = ET.SubElement(process, "bpmn:intermediateCatchEvent", {"id": node.get_id(), "name": node.get_name()})
+            messageEventDefinition = ET.SubElement(task, "bpmn:messageEventDefinition")
         elif isinstance(node, BPMN.IntermediateThrowEvent):
             task = ET.SubElement(process, "bpmn:intermediateThrowEvent", {"id": node.get_id(), "name": node.get_name()})
         elif isinstance(node, BPMN.BoundaryEvent):
             task = ET.SubElement(process, "bpmn:boundaryEvent", {"id": node.get_id(), "name": node.get_name()})
         elif isinstance(node, BPMN.Task):
-            task = ET.SubElement(process, "bpmn:task", {"id": node.get_id(), "name": node.get_name()})
+            if isinstance(node, BPMN.UserTask):
+                task = ET.SubElement(process, "bpmn:userTask", {"id": node.get_id(), "name": node.get_name()})
+            elif isinstance(node, BPMN.SendTask):
+                task = ET.SubElement(process, "bpmn:sendTask", {"id": node.get_id(), "name": node.get_name()})
+            else:
+                task = ET.SubElement(process, "bpmn:task", {"id": node.get_id(), "name": node.get_name()})
         elif isinstance(node, BPMN.SubProcess):
             task = ET.SubElement(process, "bpmn:subProcess", {"id": node.get_id(), "name": node.get_name()})
         elif isinstance(node, BPMN.ExclusiveGateway):
             task = ET.SubElement(process, "bpmn:exclusiveGateway",
                                  {"id": node.get_id(), "gatewayDirection": node.get_gateway_direction().value,
-                                  "name": ""})
+                                  "name": node.get_name()})
         elif isinstance(node, BPMN.ParallelGateway):
             task = ET.SubElement(process, "bpmn:parallelGateway",
                                  {"id": node.get_id(), "gatewayDirection": node.get_gateway_direction().value,
-                                  "name": ""})
+                                  "name": node.get_name()})
         elif isinstance(node, BPMN.InclusiveGateway):
             task = ET.SubElement(process, "bpmn:inclusiveGateway",
                                  {"id": node.get_id(), "gatewayDirection": node.get_gateway_direction().value,
-                                  "name": ""})
-        else:
-            raise Exception("Unexpected node type.")
+                                  "name": node.get_name()})
+        elif isinstance(node, BPMN.EventBasedGateway):
+            task = ET.SubElement(process, "bpmn:eventBasedGateway",
+                                 {"id": node.get_id(), "gatewayDirection": node.get_gateway_direction().value,
+                                  "name": node.get_name()})
 
         if enable_incoming_outgoing_exporting:
             for in_arc in node.get_in_arcs():
@@ -174,8 +200,20 @@ def get_xml_string(bpmn_graph, parameters=None):
 
         source = flow.get_source()
         target = flow.get_target()
-        flow_xml = ET.SubElement(process, "bpmn:sequenceFlow", {"id": "id" + str(flow.get_id()), "name": flow.get_name(),
-                                                           "sourceRef": str(source.get_id()),
-                                                           "targetRef": str(target.get_id())})
+
+        if isinstance(flow, BPMN.SequenceFlow):
+            flow_xml = ET.SubElement(process, "bpmn:sequenceFlow", {"id": "id" + str(flow.get_id()), "name": flow.get_name(),
+                                                               "sourceRef": str(source.get_id()),
+                                                               "targetRef": str(target.get_id())})
+        elif isinstance(flow, BPMN.MessageFlow):
+            flow_xml = ET.SubElement(process, "bpmn:messageFlow", {"id": "id" + str(flow.get_id()), "name": flow.get_name(),
+                                                               "sourceRef": str(source.get_id()),
+                                                               "targetRef": str(target.get_id())})
+
+
+        elif isinstance(flow, BPMN.Association):
+            flow_xml = ET.SubElement(process, "bpmn:association", {"id": "id" + str(flow.get_id()),
+                                                               "sourceRef": str(source.get_id()),
+                                                               "targetRef": str(target.get_id())})
 
     return minidom.parseString(ET.tostring(definitions)).toprettyxml(encoding=encoding)